Add organizer attribute to the Interactor perform call
This will only be set if the Interactor is being run in the context of an Organizer. If there are several layers of nested Organizers, then the organizer attribute will represent the immediate parent. If this is being run outside of an Organizer (just running the Interactor directly), organizer will be nil.