Support deferring nested after_perform calls to the end of all organized logic.
Previously, if an Organizer was nested inside another Organizer, the after_perform calls were not deferred until after the outermost Organizer had finished running...which somewhat defeats the purpose (at least when nested Organizers).

This updates the logic so that if a highly nested after_perform call is deferred, it will be deferred to the very end.